+++ es/button/index.js
@@ -20,6 +20,8 @@ export var Button = function Button(props) {
       size: props.size ? props.size : '',
       block: props.block ? 'block' : '',
       outlined: props.outlined ? 'outlined' : '',
+      allotted: props.iconAllotted ? 'allotted' : '',
+      iconLeft: props.iconLeft ? 'icon-left' : '',
       className: props.className ? props.className : ''
     };
 
@@ -31,7 +33,7 @@ export var Button = function Button(props) {
   };
 
   var setMargin = function setMargin() {
-    if (props.icon && !props.iconLeft) return 'mr-5';else if (props.icon && props.iconLeft) return 'ml-5';else return '';
+    if (props.icon && !props.iconLeft) return 'btn__text mr-5';else if (props.icon && props.iconLeft) return 'btn__text ml-5';else return 'btn__text';
   };
 
   return React.createElement("button", {
@@ -45,12 +47,22 @@ export var Button = function Button(props) {
   }, props.loading ? React.createElement(Icon, {
     name: "loading",
     color: "gray"
-  }) : React.createElement(React.Fragment, null, props.icon && props.iconLeft ? React.createElement(Icon, {
+  }) : React.createElement(React.Fragment, null, props.icon && props.iconLeft ? props.iconAllotted ? React.createElement("div", {
+    className: "btn__allotted-icon"
+  }, React.createElement(Icon, {
+    size: 20,
+    name: props.icon
+  })) : React.createElement(Icon, {
     size: 20,
     name: props.icon
   }) : '', props.name ? React.createElement("div", {
     className: setMargin()
-  }, props.name) : '', props.icon && !props.iconLeft ? React.createElement(Icon, {
+  }, props.name) : '', props.icon && !props.iconLeft ? props.iconAllotted ? React.createElement("div", {
+    className: "btn__allotted-icon"
+  }, React.createElement(Icon, {
+    size: 20,
+    name: props.icon
+  })) : React.createElement(Icon, {
     size: 20,
     name: props.icon
   }) : ''));