+++ es/pagination/index.js
@@ -70,8 +70,10 @@ export var Pagination = function Pagination(props) {
       if (currentPage % getLimit() === 0 && currentPage < allPages.length) {
         setActivePages(allPages.slice(currentPage, currentPage + getLimit()));
         setCurrentPage(currentPage + 1);
+        props.onChange(currentPage + 1);
       } else if (currentPage < allPages.length) {
         setCurrentPage(currentPage + 1);
+        props.onChange(currentPage + 1);
       }
     }
   };
@@ -81,8 +83,10 @@ export var Pagination = function Pagination(props) {
       if ((currentPage - 1) % getLimit() === 0 && currentPage > 1) {
         setActivePages(allPages.slice(currentPage - getLimit(), currentPage));
         setCurrentPage(currentPage - 1);
+        props.onChange(currentPage - 1);
       } else if (currentPage > 1) {
         setCurrentPage(currentPage - 1);
+        props.onChange(currentPage - 1);
       }
     }
   };
@@ -105,7 +109,8 @@ export var Pagination = function Pagination(props) {
       key: index,
       className: item + 1 === currentPage ? 'pagination-item active' : 'pagination-item',
       onClick: function onClick() {
-        return setCurrentPage(item + 1);
+        setCurrentPage(item + 1);
+        props.onChange(item + 1);
       }
     }, item + 1);
   }), React.createElement("div", {