+++ src/components/pagination/index.js
@@ -54,8 +54,10 @@ export const Pagination = (props) => {
             if (currentPage % getLimit() === 0 && currentPage < allPages.length) {
                 setActivePages(allPages.slice(currentPage, currentPage + getLimit()))
                 setCurrentPage(currentPage + 1)
+                props.onChange(currentPage + 1)
             } else if (currentPage < allPages.length) {
                 setCurrentPage(currentPage + 1)
+                props.onChange(currentPage + 1)
             }
        }
     }
@@ -65,8 +67,10 @@ export const Pagination = (props) => {
             if((currentPage - 1) % getLimit() === 0 && currentPage > 1){
                 setActivePages(allPages.slice(currentPage - getLimit(), currentPage))
                 setCurrentPage(currentPage - 1)
+                props.onChange(currentPage - 1)
             } else if (currentPage > 1) {
                 setCurrentPage(currentPage - 1)
+                props.onChange(currentPage - 1)
             }
         }
     }
@@ -91,7 +95,10 @@ export const Pagination = (props) => {
             {activePages.map((item, index) => {
                 return (
                     <div key={index} className={(item + 1) === currentPage ? 'pagination-item active' : 'pagination-item'}
-                        onClick={() => setCurrentPage(item + 1)}>
+                        onClick={() => {
+                            setCurrentPage(item + 1)
+                            props.onChange(item + 1)
+                        }}>
                         {item + 1}
                     </div>
                 )