+++ src/routes.js
@@ -20,6 +20,7 @@ import { PreloaderPage } from './pages/PreloaderPage';
 import { BackTopBtnPage } from './pages/BackTopBtnPage';
 import { CollapsePage } from './pages/CollapsePage';
 import { ButtonGroupPage } from './pages/ButtonGroupPage';
+import { ColorsPage } from './pages/ColorsPage';
 
 export const routes = [
     { path: '/', name: 'Get started', Component: HomePage, icon: 'rocket' },
@@ -43,5 +44,6 @@ export const routes = [
     { path: '/preloader', name: 'Preloader', Component: PreloaderPage },
     { path: '/backtopbtn', name: 'BackTopBtn', Component: BackTopBtnPage },
     { path: '/collapse', name: 'Collapse', Component: CollapsePage },
-    { path: '/buttonGroup', name: 'ButtonGroup', Component: ButtonGroupPage }
+    { path: '/buttonGroup', name: 'ButtonGroup', Component: ButtonGroupPage },
+    { path: '/colors', name: 'Colors', Component: ColorsPage }
 ]