+++ CHANGELOG.md
@@ -7,6 +7,13 @@ timeline: true
 
 ###### For now there is not schedule, make releases as soon as possible and when i have a time
 
+#### 0.2.24 `2020-02-20`
+
+- Button component updated;
+- Table component updated;
+- Pagination component updated;
+- `NEW!` New icons added;
+
 #### 0.2.23 `2020-02-19`
 
 - `NEW!` ButtonGroup component added;